+++ rules/react.js
@@ -1,12 +1,19 @@
 /* eslint-env node */
 
 module.exports = {
-    plugins: [ "react" ],
+    plugins: [
+        "react",
+        "jsx-control-statements",
+    ],
 
     parserOptions: {
         ecmaFeatures: {
             jsx: true,
         },
+    },
+
+    env: {
+        "jsx-control-statements/jsx-control-statements": true,
     },
 
     rules: {
@@ -26,7 +33,7 @@ module.exports = {
         "react/jsx-no-target-blank": "error",
 
         // Disallow undeclared variables in JSX
-        "react/jsx-no-undef": "error",
+        "react/jsx-no-undef": [ "error", { allowGlobals: true } ],
 
         // Prevent React to be incorrectly marked as unused
         "react/jsx-uses-react": "error",
@@ -91,5 +98,56 @@ module.exports = {
         // Ensure no casing typos were made declaring static class properties
         // and lifecycle methods.
         "react/no-typos": "error",
+
+        /*
+         * Choose tag is empty or does not have at least one When
+         * tag as child.
+         */
+        "jsx-control-statements/jsx-choose-not-empty": "error",
+
+        /*
+         * Warn if For tag is missing each attribute. And also marks the
+         * variable as defined.
+         */
+        "jsx-control-statements/jsx-for-require-each": "error",
+
+        /*
+         * Warn if For tag is missing of attribute.
+         */
+        "jsx-control-statements/jsx-for-require-of": "error",
+
+        /*
+         * Warn if If tag is missing condition attribute.
+         */
+        "jsx-control-statements/jsx-if-require-condition": "error",
+
+        /*
+         * Warn when Otherwise tag is used more than once inside Choose and is
+         * not last child.
+         */
+        "jsx-control-statements/jsx-otherwise-once-last": "error",
+
+        /*
+         * Use If tag instead of ternary operator.
+         */
+        "jsx-control-statements/jsx-use-if-tag": "error",
+
+        /*
+         * Warn if When tag is missing condition attribute.
+         */
+        "jsx-control-statements/jsx-when-require-condition": "error",
+
+        /*
+         * This rule is the same as the generic eslint no-undef rule
+         * (see http://eslint.org/docs/rules/no-undef) except with an
+         * exception built in for variables that are implicitly declared by
+         * <For> and <With> statements. Note that this includes no-undef's
+         * code and completely replaces it rather than supplementing it - if
+         * this rule is on, no-undef should be off. It is compatible with
+         * no-undef's options and global declarations.
+         */
+        "jsx-control-statements/jsx-jcs-no-undef": "error",
+
+        "no-undef": "off",
     },
 }